+++ spec/ext_spec.rb
@@ -1,20 +1,8 @@
-# -*- coding: ascii
-require 'rmodbus'
-
 describe Array do
   before do
-    @arr = [1,0,1,1, 0,0,1,1, 1,1,0,1, 0,1,1,0,  1,0,1]
     @test = [0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0]
   end
 
-  it "should return string reprisent 16bit" do
-    @arr.pack_to_word.should == "\xcd\x6b\x5" 
-  end
-
-  it "fixed bug for	divisible 8 data " do
-    ([0] * 8).pack_to_word.should == "\x00"
-  end
-  
   it "should unpack to @test" do
     "test".unpack_bits == @test
   end
@@ -24,7 +12,7 @@ describe Array do
     [20342, 17344, 20342, 17344].to_32i.size.should == 2
   end
 
-  it "should turn an array into 32b floats" do
+  it "should turn an array into 32b floats, big endian" do
     [20342, 17344].to_32f[0].should be_within(0.1).of(384.620788574219)
     [20342, 17344, 20342, 17344].to_32f.size.should == 2
   end
@@ -44,4 +32,3 @@ describe Array do
    lambda { [20342, 17344, 123].to_32i }.should raise_error(StandardError)
   end
 end
-